refactor: simplify artist profile page data flow
