+++ src/api/account/getCurrentAccount.ts
@@ -3,10 +3,8 @@ import { CurrentAccountResponse } from "./types";
 
 const CURRENT_ACCOUNT_PATH = "/api/account/me";
 
-export async function getCurrentAccount(
-  token: string,
-): Promise<CurrentAccountResponse> {
-  return requestAccount<CurrentAccountResponse>(CURRENT_ACCOUNT_PATH, token, {
+export async function getCurrentAccount(): Promise<CurrentAccountResponse> {
+  return requestAccount<CurrentAccountResponse>(CURRENT_ACCOUNT_PATH, {
     method: "GET",
   });
 }
