+++ src/api/account/updateAccountPhone.ts
@@ -5,11 +5,9 @@ const CURRENT_ACCOUNT_CHANGE_PHONE_PATH = "/api/account/me/change-phone";
 
 export async function updateAccountPhone(
   payload: UpdateAccountPhonePayload,
-  token: string,
 ): Promise<UpdateAccountPhoneResponse> {
   return requestAccount<UpdateAccountPhoneResponse>(
     CURRENT_ACCOUNT_CHANGE_PHONE_PATH,
-    token,
     {
       method: "PATCH",
       headers: {
