+++ src/api/account/updateAccountUsername.ts
@@ -8,11 +8,9 @@ const CURRENT_ACCOUNT_CHANGE_USERNAME_PATH = "/api/account/me/change-username";
 
 export async function updateAccountUsername(
   payload: UpdateAccountUsernamePayload,
-  token: string,
 ): Promise<UpdateAccountUsernameResponse> {
   return requestAccount<UpdateAccountUsernameResponse>(
     CURRENT_ACCOUNT_CHANGE_USERNAME_PATH,
-    token,
     {
       method: "PATCH",
       headers: {
