+++ src/api/artist/getCurrentArtist.ts
@@ -3,10 +3,8 @@ import { CurrentArtistResponse } from "./types";
 
 const CURRENT_ARTIST_PATH = "/api/artist/me";
 
-export async function getCurrentArtist(
-  token: string,
-): Promise<CurrentArtistResponse> {
-  return requestArtist<CurrentArtistResponse>(CURRENT_ARTIST_PATH, token, {
+export async function getCurrentArtist(): Promise<CurrentArtistResponse> {
+  return requestArtist<CurrentArtistResponse>(CURRENT_ARTIST_PATH, {
     method: "GET",
   });
 }
