+++ src/api/artist/updateCurrentArtist.ts
@@ -8,19 +8,14 @@ const CURRENT_ARTIST_PATH = "/api/artist/me";
 
 export async function updateCurrentArtist(
   payload: UpdateCurrentArtistPayload,
-  token: string,
 ): Promise<UpdateCurrentArtistResponse> {
-  return requestArtist<UpdateCurrentArtistResponse>(
-    CURRENT_ARTIST_PATH,
-    token,
-    {
-      method: "PATCH",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(payload),
+  return requestArtist<UpdateCurrentArtistResponse>(CURRENT_ARTIST_PATH, {
+    method: "PATCH",
+    headers: {
+      "Content-Type": "application/json",
     },
-  );
+    body: JSON.stringify(payload),
+  });
 }
 
 export default updateCurrentArtist;
