+++ src/api/artist/updateCurrentArtistCover.ts
@@ -8,14 +8,12 @@ const CURRENT_ARTIST_COVER_PATH = "/api/artist/me/cover";
 
 export async function updateCurrentArtistCover(
   payload: UpdateCurrentArtistCoverPayload,
-  token: string,
 ): Promise<UpdateCurrentArtistCoverResponse> {
   const formData = new FormData();
   formData.set("cover", payload.cover);
 
   return requestArtist<UpdateCurrentArtistCoverResponse>(
     CURRENT_ARTIST_COVER_PATH,
-    token,
     {
       method: "PATCH",
       body: formData,
